Refactoring: Move asserthandler header to include
