+++ code/Common/AssertHandler.cpp
@@ -43,7 +43,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  *  @brief Implementation of assert handling logic.
  */
 
-#include "AssertHandler.h"
+#include <assimp/AssertHandler.h>
 
 #include <iostream>
 #include <cstdlib>
