+++ include/assimp/AssertHandler.h
