+++ test/unit/Common/utBaseProcess.cpp
@@ -43,7 +43,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include "UnitTestPCH.h"
 
 #include "Common/BaseProcess.h"
-#include "Common/AssertHandler.h"
+#include <assimp/AssertHandler.h>
 
 using namespace Assimp;
